# docs/atlas-prd/03-contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Atlas Wallet and Financial Operations API
  version: 2026-07-23
  summary: Security-first contracts for wallet, transfers, merchant payments, operations, and reporting.
  description: |
    Reference contract for the Atlas portfolio platform. Financial amounts are integer minor-unit
    strings. Money-moving commands are idempotent. Browser applications use a backend-for-frontend
    session; merchant and service integrations use scoped machine credentials. This file is a
    deliberately curated baseline, not an automatically inferred description of handlers.
  contact:
    name: Atlas API Governance
  license:
    name: Portfolio reference - no production authorization
servers:
  - url: https://api.staging.atlas.example
    description: Synthetic staging environment
  - url: http://localhost:8080
    description: Local development
...
security:
  - bffSession: []
paths:
...
  /v1/me:
    get:
      operationId: getCurrentPrincipal
      tags: [Identity]
      summary: Get the authenticated principal and active tenant context
      description: An invitation-acceptance session returns a null active tenant and no permissions until acceptance commits.
      security:
        - bffSession: []
        - merchantApiKey: []
      responses:
        '200':
          description: Current principal
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            X-Atlas-CSRF-Token: {$ref: '#/components/headers/XAtlasCSRFToken'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Principal'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/auth/login:
    get:
      operationId: beginBrowserLogin
      tags: [Identity]
      summary: Begin a synthetic OIDC authorization-code login
      description: Creates a short-lived single-use login transaction and redirects only to an allow-listed issuer.
      security: []
      parameters:
        - $ref: '#/components/parameters/AuthPopulation'
        - $ref: '#/components/parameters/ReturnTo'
      responses:
        '303':
          description: Redirect to the allow-listed synthetic identity provider
          headers:
            Location: {$ref: '#/components/headers/Location'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/auth/callback:
    get:
      operationId: completeBrowserLogin
      tags: [Identity]
      summary: Complete a synthetic OIDC authorization-code login
      description: Validates issuer, state, nonce, PKCE, code redemption, and claim mapping before issuing an opaque BFF session.
      security: []
      parameters:
        - $ref: '#/components/parameters/AuthorizationCode'
        - $ref: '#/components/parameters/OIDCState'
        - $ref: '#/components/parameters/OIDCAuthorizationIssuer'
        - $ref: '#/components/parameters/OIDCSessionState'
      responses:
        '303':
          description: Redirect to a same-origin allow-listed application path
          headers:
            Location: {$ref: '#/components/headers/Location'}
            Set-Cookie: {$ref: '#/components/headers/SessionSetCookie'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/logout:
    post:
      operationId: logoutCurrentSession
      tags: [Identity]
      summary: Revoke the current browser session
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
      responses:
        '204':
          description: Current session revoked or already absent
          headers:
            Set-Cookie: {$ref: '#/components/headers/SessionSetCookie'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/sessions:
    get:
      operationId: listPrincipalSessions
      tags: [Identity]
      summary: List sessions owned by the current principal
      responses:
        '200':
          description: Current principal session list
          headers:
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data]
                properties:
                  data:
                    type: array
                    maxItems: 100
                    items: {$ref: '#/components/schemas/Session'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
  /v1/sessions/{session_id}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    delete:
      operationId: revokePrincipalSession
      tags: [Identity]
      summary: Revoke one session owned by the current principal
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
      responses:
        '204': {description: Session revoked or already absent}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
  /v1/sessions/revoke-all:
    post:
      operationId: revokeAllPrincipalSessions
      tags: [Identity]
      summary: Revoke all sessions for the current principal
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RevokeAllSessionsRequest'}
      responses:
        '204': {description: Selected sessions revoked}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/security/sessions/{session_id}/revocations:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      operationId: revokeSessionForSecurity
      tags: [Identity]
      summary: Revoke one application session under workforce security authority
      description: |
        Requires the `identity.sessions.revoke_admin` permission, the exact
        `security_review` purpose, and a fresh phishing-resistant step-up bound to
        `identity.session.admin_revoke`. The authorization decision, revocation,
        idempotency result, and Audit fact commit atomically in PostgreSQL. Target
        absence is concealed until the workforce authorization check succeeds.
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminSessionRevocationRequest'}
      responses:
        '204':
          description: Session revoked or already revoked
          headers:
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403':
          description: Permission is denied or a fresh action-bound step-up is required
          headers:
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/Problem'}
        '404':
          description: Session is absent or concealed after workforce authorization succeeds
          headers:
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/Problem'}
        '409': {$ref: '#/components/responses/Conflict'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/step-up/challenges:
    post:
      operationId: createStepUpChallenge
      tags: [Identity]
      summary: Begin a fresh higher-assurance authentication challenge
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StepUpChallengeRequest'}
      responses:
        '201':
          description: Step-up challenge created
          headers:
            Location: {$ref: '#/components/headers/Location'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StepUpChallenge'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/me/active-organization:
    put:
      operationId: setActiveOrganization
      tags: [Organizations]
      summary: Select the current browser session organization context
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SetActiveOrganizationRequest'}
      responses:
        '200':
          description: Updated principal and tenant context
          headers:
            Set-Cookie: {$ref: '#/components/headers/SessionSetCookie'}
            X-Atlas-CSRF-Token: {$ref: '#/components/headers/XAtlasCSRFToken'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Principal'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
  /v1/organizations:
    get:
      operationId: listPrincipalOrganizations
      tags: [Organizations]
      summary: List organizations visible to the current principal
      responses:
        '200':
          description: Organization list
          headers:
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data]
                properties:
                  data:
                    type: array
                    maxItems: 100
                    items: {$ref: '#/components/schemas/Organization'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/organizations/{organization_id}/members:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      operationId: listOrganizationMembers
      tags: [Organizations]
      summary: List authorized members without exposing concealed membership
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/XAtlasPurpose'
      responses:
        '200':
          description: Authorized organization-member page
          headers:
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, page]
                properties:
                  data:
                    type: array
                    maxItems: 100
                    items: {$ref: '#/components/schemas/OrganizationMember'}
                  page: {$ref: '#/components/schemas/PageInfo'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
  /v1/organizations/{organization_id}/invitations:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    post:
      operationId: createOrganizationInvitation
      tags: [Organizations]
      summary: Create a single-use organization invitation
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateOrganizationInvitationRequest'}
      responses:
        '201':
          description: Invitation created; the acceptance token is disclosed only for the original response
          headers:
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Location: {$ref: '#/components/headers/Location'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OrganizationInvitationCreated'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /v1/organization-invitations/{invitation_id}/authentication:
    parameters:
      - $ref: '#/components/parameters/InvitationId'
    post:
      operationId: beginOrganizationInvitationAuthentication
      tags: [Organizations]
      summary: Begin recipient-bound merchant authentication for an invitation
      description: Accepts the token only in the body, stores only its digest in a single-use OIDC transaction, and does not reveal whether the invitation exists before the verified callback.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AcceptOrganizationInvitationRequest'}
      responses:
        '303':
          description: Redirect to the allow-listed synthetic merchant identity provider
          headers:
            Location: {$ref: '#/components/headers/Location'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/organization-invitations/{invitation_id}/acceptance:
    parameters:
      - $ref: '#/components/parameters/InvitationId'
    post:
      operationId: acceptOrganizationInvitation
      tags: [Organizations]
      summary: Accept a single-use organization invitation
      description: Requires the recipient-bound acceptance-only BFF session created by the authentication flow; membership, invitation consumption, session rotation, and Audit commit atomically.
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AcceptOrganizationInvitationRequest'}
      responses:
        '200':
          description: Invitation accepted and membership returned
          headers:
            Set-Cookie: {$ref: '#/components/headers/SessionSetCookie'}
            X-Atlas-CSRF-Token: {$ref: '#/components/headers/XAtlasCSRFToken'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OrganizationMember'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /v1/organizations/{organization_id}/members/{member_id}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - $ref: '#/components/parameters/MemberId'
    patch:
      operationId: updateOrganizationMember
      tags: [Organizations]
      summary: Change an organization member role subject to approval policy
      description: Direct changes are limited to merchant_viewer and merchant_operator. Any transition into or out of merchant_admin is the typed maker-checker action owned by Operations and returns 202 only after an approval is durably created; merchant_security_admin is never delegable through this operation. If-Match is the exact strong membership ETag returned by Atlas.
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateOrganizationMemberRequest'}
      responses:
        '200':
          description: Membership changed
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OrganizationMember'}
        '202':
          description: Approval is required before execution
          headers:
            Location: {$ref: '#/components/headers/Location'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
    delete:
      operationId: revokeOrganizationMember
      tags: [Organizations]
      summary: Revoke organization membership subject to last-admin policy
      description: Direct revocation is currently limited to merchant_viewer and merchant_operator memberships. Administrator targets remain fail-closed until their fresh-step-up and last-administrator execution policy is implemented.
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Membership revoked or an identical committed revocation replayed
          headers:
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
  /v1/approvals:
    get:
      operationId: listApprovals
      tags: [Approvals]
      summary: List approvals visible to the active tenant and principal
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Approval page
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/Approval'}
                  page: {$ref: '#/components/schemas/PageInfo'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
    post:
      operationId: createApproval
      tags: [Approvals]
      summary: Create a typed approval request over a canonical payload digest
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateApprovalRequest'}
      responses:
        '201':
          description: Approval request created
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Location: {$ref: '#/components/headers/Location'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/StepUpRequired'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/approvals/{approval_id}:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    get:
      operationId: getApproval
      tags: [Approvals]
      summary: Get an approval and its immutable payload digest
      responses:
        '200':
          description: Approval detail
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
components:
  securitySchemes:
    bffSession:
      type: apiKey
      in: cookie
      name: __Host-atlas_session
      description: Secure, HttpOnly, SameSite session cookie managed by the BFF.
    merchantApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: |
        AtlasKey credential formatted as `AtlasKey <key-id>.<secret>`. Secrets are high-entropy,
        one-time-disclosed values stored only as SHA-256 verifiers. The `identity:read` scope
        authorizes only `GET /v1/me` in Phase 01.
    merchantOAuth:
      type: oauth2
      description: OAuth 2.0 client-credentials access token with audience and scope enforcement.
      flows:
        clientCredentials:
          tokenUrl: https://identity.staging.atlas.example/oauth2/token
          scopes:
            payments:read: Read payment intents
            payments:write: Create and capture payment intents
            refunds:read: Read refunds
            refunds:write: Create refunds
            webhooks:read: Read webhook endpoints and deliveries
            webhooks:write: Manage endpoints and replay deliveries
...
  headers:
    XRequestId:
      schema: {$ref: '#/components/schemas/OpaqueId'}
...
    CacheControl:
      schema: {type: string, const: no-store}
    ETag:
      schema: {type: string}
...
    XAtlasCSRFToken:
      description: New memory-only CSRF token bound to the current session and latest rotation.
      schema:
        type: string
        minLength: 43
        maxLength: 128
        pattern: '^[A-Za-z0-9_-]+$'
...